% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Beta_est_sim}
\alias{Beta_est_sim}
\title{####################################################################
#####the function is for log likelihood function of the data########
\subsection{function argurments:}{

##b1: the parameter for x1 in the hazard funtion for the terminal event;
##b2: the parameter for x2 in the hazard funtion for the terminal event;
##b3: the parameter for x1 in the hazard funtion for the recurrent event;
##b4: the parameter for x2 in the hazard funtion for the recurrent event;
##b01, b02: baseline hazards for the terminal event and the recurrent events, respectively;
##phi: random effect term denoted as w in paper; {W_i}
##theta: correlation parameter for copula function; {theta_ij}
}

\subsection{Mu: theta_mu average level correlation}{

##x1, x2: two covariates in the hazard function for the terminal event and the recurrent event;
##y1, y2: observed time for the terminal event and the recurrent event;
##s1, s2: failing indicator for the terminal event and the recurrent event;
##id: subject id;
##N: the sample size of subjects;
b01_est1,b02_est1,b1_est1,b2_est1,b3_est1,b4_est1: coefficients in JHCM
b01_est,b02_est,b1_est,b2_est,b3_est,b4_est: Coefficiets in TVJHCM
####################################################################
##########    usage ##############################
idx<-4
numCores <- 4
registerDoParallel(numCores)
nsim<-4
result<- foreach(n.sim=1:nsim,.errorhandling = "remove") \%dopar\% {
Beta_est_sim(idx,n.sim)
}
saveRDS(result,paste0("400_WSF_old_",idx,".rds"))
}}
\usage{
Beta_est_sim(idx, n.sim)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
estimation for coefficient in simulation
}
\description{
####################################################################
#####the function is for log likelihood function of the data########
\subsection{function argurments:}{

##b1: the parameter for x1 in the hazard funtion for the terminal event;
##b2: the parameter for x2 in the hazard funtion for the terminal event;
##b3: the parameter for x1 in the hazard funtion for the recurrent event;
##b4: the parameter for x2 in the hazard funtion for the recurrent event;
##b01, b02: baseline hazards for the terminal event and the recurrent events, respectively;
##phi: random effect term denoted as w in paper; {W_i}
##theta: correlation parameter for copula function; {theta_ij}
}

\subsection{Mu: theta_mu average level correlation}{

##x1, x2: two covariates in the hazard function for the terminal event and the recurrent event;
##y1, y2: observed time for the terminal event and the recurrent event;
##s1, s2: failing indicator for the terminal event and the recurrent event;
##id: subject id;
##N: the sample size of subjects;
b01_est1,b02_est1,b1_est1,b2_est1,b3_est1,b4_est1: coefficients in JHCM
b01_est,b02_est,b1_est,b2_est,b3_est,b4_est: Coefficiets in TVJHCM
####################################################################
##########    usage ##############################
idx<-4
numCores <- 4
registerDoParallel(numCores)
nsim<-4
result<- foreach(n.sim=1:nsim,.errorhandling = "remove") \%dopar\% {
Beta_est_sim(idx,n.sim)
}
saveRDS(result,paste0("400_WSF_old_",idx,".rds"))
}
}
