% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS_WSFM.R
\name{BS_figure}
\alias{BS_figure}
\title{####################################################################
########### This function  return figure of BS #############
############ input #########
idx: index control simulation settings
##########  Output  ############
figure for BS
####################################################################
##########    usage ##############################
result<-readRDS("200_WSF_72.rds")
BS(72)}
\usage{
BS_figure(idx)
}
\value{
graph for BS
}
\description{
####################################################################
########### This function  return figure of BS #############
############ input #########
idx: index control simulation settings
##########  Output  ############
figure for BS
####################################################################
##########    usage ##############################
result<-readRDS("200_WSF_72.rds")
BS(72)
}
